collision corrected

+++ Engine/CollisionComponent.h
@@ -3,10 +3,12 @@
 class CollisionComponent : public Component
 {
 public:
-	CollisionComponent(int x, int y);
+	CollisionComponent(float x, float y);
 
 	void Update(float dt) override;
 	void Render(sf::RenderWindow* w) override;
+
+
 private:
 	sf::RectangleShape rect;
 };
